% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{CountryPopulation}
\usage{
CountryPopulation(Country.Name)
}
\arguments{
\item{Country.Name}{country within the WorldPopulation dataset}
}
\value{
A population vs year bar graph for the specific country that is input into the function.
}
\description{
This function produces a population over time bar graph for a specific country
that is in the WorldPopulation dataset.
}
\examples{
CountryPopulation("South Africa")
CountryPopulation("Dominica")
CountryPopulation("Finland")

}
